Add movie rating average

+++ src/main/java/com/example/gmovie/model/Movie.java
@@ -20,7 +20,7 @@ public class Movie {
 
     private String title;
 
-    private int rating;
+    private float rating;
 
 
 }
